refactor(plugin): extract PluginFormat strategy for plugin loading
Move on-disk plugin parsing out of Plugin into a PluginFormat strategy.
Plugin.load() now detects the format via detect_format() and delegates
to the strategy, which returns a normalized Plugin. The Claude Code
layout is preserved verbatim as ClaudeCodePluginFormat; skills discovery
and final assembly are shared on the base class.

No behavior change. This carves the seam for adding the Agent Plugins
(agent-plugins.org) format without touching the merge/apply path.
