+++ openhands-sdk/openhands/sdk/plugin/__init__.py
@@ -21,6 +21,11 @@ from openhands.sdk.plugin.fetch import (
     PluginFetchError,
     fetch_plugin_with_resolution,
 )
+from openhands.sdk.plugin.format import (
+    ClaudeCodePluginFormat,
+    PluginFormat,
+    detect_format,
+)
 from openhands.sdk.plugin.installed import (
     InstalledPluginInfo,
     disable_plugin,
@@ -56,6 +61,10 @@ __all__ = [
     "Plugin",
     "PluginFetchError",
     "PluginManifest",
+    # Plugin format strategies
+    "PluginFormat",
+    "ClaudeCodePluginFormat",
+    "detect_format",
     "PluginAuthor",
     "PluginSource",
     "ResolvedPluginSource",
