+++ openhands-sdk/openhands/sdk/plugin/plugin.py
@@ -1,8 +1,14 @@
-"""Plugin class for loading and managing plugins."""
+"""Plugin class for loading and managing plugins.
+
+``Plugin`` is the format-neutral, in-memory model plus its format-agnostic
+behavior (merging skills / MCP config into an agent). Reading a plugin directory
+off disk is delegated to a :class:`~openhands.sdk.plugin.format.PluginFormat`
+strategy (see ``format.py``); ``Plugin.load()`` is a thin dispatcher that detects
+the on-disk format and returns a normalized ``Plugin``.
+"""
 
 from __future__ import annotations
 
-import json
 from pathlib import Path
 from typing import TYPE_CHECKING
 
@@ -10,30 +16,20 @@ from pydantic import BaseModel, Field
 
 from openhands.sdk.hooks import HookConfig
 from openhands.sdk.logger import get_logger
-from openhands.sdk.mcp.config import MCPServer, coerce_mcp_config
+from openhands.sdk.mcp.config import MCPServer
 from openhands.sdk.plugin.fetch import fetch_plugin
 from openhands.sdk.plugin.types import (
     CommandDefinition,
-    PluginAuthor,
     PluginManifest,
 )
 from openhands.sdk.skills.skill import Skill
-from openhands.sdk.skills.utils import (
-    find_skill_md,
-    load_mcp_config,
-)
 from openhands.sdk.subagent.schema import AgentDefinition
-from openhands.sdk.utils.path import to_posix_path
 
 
 if TYPE_CHECKING:
     from openhands.sdk.context import AgentContext
 
 logger = get_logger(__name__)
-
-# Directories to check for plugin manifest
-PLUGIN_MANIFEST_DIRS = [".plugin", ".claude-plugin"]
-PLUGIN_MANIFEST_FILE = "plugin.json"
 
 
 class Plugin(BaseModel):
@@ -273,37 +269,15 @@ class Plugin(BaseModel):
             FileNotFoundError: If the plugin directory doesn't exist.
             ValueError: If the plugin manifest is invalid.
         """
+        # Imported here (not at module scope) because format.py imports Plugin
+        # for its return type; a top-level import would create a cycle.
+        from openhands.sdk.plugin.format import detect_format
+
         plugin_dir = Path(plugin_path).resolve()
         if not plugin_dir.is_dir():
             raise FileNotFoundError(f"Plugin directory not found: {plugin_dir}")
 
-        # Load manifest
-        manifest = _load_manifest(plugin_dir)
-
-        # Load skills
-        skills = _load_skills(plugin_dir)
-
-        # Load hooks
-        hooks = _load_hooks(plugin_dir)
-
-        # Load MCP config
-        mcp_config = _load_plugin_mcp_config(plugin_dir)
-
-        # Load agents
-        agents = _load_agents(plugin_dir)
-
-        # Load commands
-        commands = _load_commands(plugin_dir)
-
-        return cls(
-            manifest=manifest,
-            path=to_posix_path(plugin_dir),
-            skills=skills,
-            hooks=hooks,
-            mcp_config=mcp_config,
-            agents=agents,
-            commands=commands,
-        )
+        return detect_format(plugin_dir).load(plugin_dir)
 
     @classmethod
     def load_all(cls, plugins_dir: str | Path) -> list[Plugin]:
@@ -331,205 +305,3 @@ class Plugin(BaseModel):
                     logger.warning(f"Failed to load plugin from {item}: {e}")
 
         return plugins
-
-
-def _load_manifest(plugin_dir: Path) -> PluginManifest:
-    """Load plugin manifest from plugin.json.
-
-    Checks both .plugin/ and .claude-plugin/ directories.
-    Falls back to inferring from directory name if no manifest found.
-    """
-    manifest_path = None
-
-    # Check for manifest in standard locations
-    for manifest_dir in PLUGIN_MANIFEST_DIRS:
-        candidate = plugin_dir / manifest_dir / PLUGIN_MANIFEST_FILE
-        if candidate.exists():
-            manifest_path = candidate
-            break
-
-    if manifest_path:
-        try:
-            with open(manifest_path, encoding="utf-8") as f:
-                data = json.load(f)
-
-            # Handle author field - can be string or object
-            if "author" in data and isinstance(data["author"], str):
-                data["author"] = PluginAuthor.from_string(data["author"]).model_dump()
-
-            return PluginManifest.model_validate(data)
-        except json.JSONDecodeError as e:
-            raise ValueError(f"Invalid JSON in {manifest_path}: {e}") from e
-        except Exception as e:
-            raise ValueError(f"Failed to parse manifest {manifest_path}: {e}") from e
-
-    # Fall back to inferring from directory name
-    logger.debug(f"No manifest found for {plugin_dir}, inferring from directory name")
-    return PluginManifest(
-        name=plugin_dir.name,
-        version="1.0.0",
-        description=f"Plugin loaded from {plugin_dir.name}",
-    )
-
-
-def _load_skills(plugin_dir: Path) -> list[Skill]:
-    """Load a plugin's skills.
-
-    Supports both Claude Code plugin skill layouts:
-
-    - Multi-skill: a ``skills/`` directory containing one ``<name>/SKILL.md``
-      per skill (or single ``.md`` files).
-    - Single-skill: a ``SKILL.md`` at the plugin root when there is no
-      ``skills/`` directory. Claude Code loads such a plugin as a single-skill
-      plugin (v2.1.142+); this mirrors that behavior so standalone Agent Skills
-      published as plugins load without an extra nesting level.
-
-    Note: Plugin skills are loaded with relaxed validation (strict=False)
-    to support Claude Code plugins which may use different naming conventions.
-    """
-    skills_dir = plugin_dir / "skills"
-    if skills_dir.is_dir():
-        return _load_skills_from_skills_dir(skills_dir)
-
-    root_skill_md = find_skill_md(plugin_dir)
-    if root_skill_md is not None:
-        return _load_root_skill(plugin_dir, root_skill_md)
-
-    return []
-
-
-def _load_skills_from_skills_dir(skills_dir: Path) -> list[Skill]:
-    """Load every skill under a plugin's ``skills/`` directory."""
-    skills: list[Skill] = []
-    for item in sorted(skills_dir.iterdir()):
-        if item.is_dir():
-            skill_md = find_skill_md(item)
-            if skill_md:
-                try:
-                    # Skill.load() discovers resources, no need to do it again
-                    skill = Skill.load(skill_md, skills_dir, strict=False)
-                    skills.append(skill)
-                    logger.debug(f"Loaded skill: {skill.name} from {skill_md}")
-                except Exception as e:
-                    logger.warning(f"Failed to load skill from {item}: {e}")
-        elif item.suffix == ".md" and item.name.lower() != "readme.md":
-            # Also support single .md files in skills/ directory
-            try:
-                skill = Skill.load(item, skills_dir, strict=False)
-                skills.append(skill)
-                logger.debug(f"Loaded skill: {skill.name} from {item}")
-            except Exception as e:
-                logger.warning(f"Failed to load skill from {item}: {e}")
-
-    return skills
-
-
-def _load_root_skill(plugin_dir: Path, skill_md: Path) -> list[Skill]:
-    """Load a single-skill plugin whose ``SKILL.md`` lives at the plugin root.
-
-    For root skills, the plugin directory is the skill root, so .mcp.json at the
-    plugin level is the same file that Skill.load() would try to load. We pass
-    skip_mcp=True to avoid double-loading with different semantics (plugin-level
-    uses expand_defaults=False for deferred secret expansion; skill-level would
-    use expand_defaults=True and raise on validation errors).
-    """
-    try:
-        # skip_mcp=True: Plugin-level MCP already loaded
-        # Skill.load() discovers resources, no need to do it again
-        skill = Skill.load(skill_md, plugin_dir, strict=False, skip_mcp=True)
-        logger.debug(f"Loaded single-skill plugin: {skill.name} from {skill_md}")
-        return [skill]
-    except Exception as e:
-        logger.warning(f"Failed to load root skill from {plugin_dir}: {e}")
-        return []
-
-
-def _load_hooks(plugin_dir: Path) -> HookConfig | None:
-    """Load hooks configuration from hooks/hooks.json."""
-    hooks_json = plugin_dir / "hooks" / "hooks.json"
-    if not hooks_json.exists():
-        return None
-
-    try:
-        hook_config = HookConfig.load(path=hooks_json)
-        # If hooks.json exists but is invalid, HookConfig.load() returns an empty
-        # config and logs the validation error. Keep that distinct from "file not
-        # present" (None).
-        if hook_config.is_empty():
-            logger.info(f"No hooks configured in {hooks_json}")
-            return HookConfig()
-        logger.info(f"Loaded hooks from {hooks_json}")
-        return hook_config
-    except Exception as e:
-        logger.warning(f"Failed to load hooks from {hooks_json}: {e}")
-        return None
-
-
-def _load_plugin_mcp_config(plugin_dir: Path) -> dict[str, MCPServer]:
-    """Load MCP config from .mcp.json.
-
-    Note: Variables are NOT fully expanded during plugin loading. Only SKILL_ROOT
-    is expanded (since plugin_dir is known). Other variables like ${VAR:-default}
-    are preserved as placeholders to be expanded later when per-conversation
-    secrets are available (in LocalConversation._ensure_plugins_loaded()).
-
-    This prevents the double-expansion bug where defaults would be applied
-    during plugin loading before secrets are available.
-    """
-    mcp_json = plugin_dir / ".mcp.json"
-    if not mcp_json.exists():
-        return {}
-
-    try:
-        # expand_defaults=False: preserve ${VAR:-default} placeholders for later
-        # expansion with per-conversation secrets. Only SKILL_ROOT is expanded now.
-        config = load_mcp_config(mcp_json, skill_root=plugin_dir, expand_defaults=False)
-        if config and "mcpServers" in config:
-            logger.info(
-                "Loaded MCP config from %s with %d server(s)",
-                mcp_json,
-                len(config["mcpServers"]),
-            )
-        servers = config.get("mcpServers", {}) if isinstance(config, dict) else {}
-        return coerce_mcp_config(servers)
-    except Exception as e:
-        logger.warning(f"Failed to load MCP config from {mcp_json}: {e}")
-        return {}
-
-
-def _load_agents(plugin_dir: Path) -> list[AgentDefinition]:
-    """Load agent definitions from the agents/ directory."""
-    agents_dir = plugin_dir / "agents"
-    if not agents_dir.is_dir():
-        return []
-
-    agents: list[AgentDefinition] = []
-    for item in sorted(agents_dir.iterdir()):
-        if item.suffix == ".md" and item.name.lower() != "readme.md":
-            try:
-                agent = AgentDefinition.load(item)
-                agents.append(agent)
-                logger.debug(f"Loaded agent: {agent.name} from {item}")
-            except Exception as e:
-                logger.warning(f"Failed to load agent from {item}: {e}")
-
-    return agents
-
-
-def _load_commands(plugin_dir: Path) -> list[CommandDefinition]:
-    """Load command definitions from the commands/ directory."""
-    commands_dir = plugin_dir / "commands"
-    if not commands_dir.is_dir():
-        return []
-
-    commands: list[CommandDefinition] = []
-    for item in sorted(commands_dir.iterdir()):
-        if item.suffix == ".md" and item.name.lower() != "readme.md":
-            try:
-                command = CommandDefinition.load(item)
-                commands.append(command)
-                logger.debug(f"Loaded command: {command.name} from {item}")
-            except Exception as e:
-                logger.warning(f"Failed to load command from {item}: {e}")
-
-    return commands
